+++ caravel/caravel.py
@@ -4,6 +4,7 @@ from functools import wraps
 import getpass
 import traceback
 import warnings
+from yaml import safe_load
 from flask import Flask, render_template, request, jsonify, session, redirect, send_from_directory, url_for, flash
 import globs
 from .const import *
@@ -105,9 +106,9 @@ def generate_csrf_token(n=100):
     """
     if '_csrf_token' not in session:
         session['_csrf_token'] = random_string(n)
-        app.logger.info("CSRF token generated")
+        app.logger.debug("CSRF token generated")
     else:
-        app.logger.info("CSRF token retrieved from the session")
+        app.logger.debug("CSRF token retrieved from the session")
     return session['_csrf_token']
 
 
@@ -157,7 +158,7 @@ def parse_token_file(path=TOKEN_FILE_NAME):
     """
     try:
         with open(path, 'r') as stream:
-            out = yaml.safe_load(stream)
+            out = safe_load(stream)
         assert CONFIG_TOKEN_KEY in out, \
             "'{token}' key not in the {file} file.".format(token=CONFIG_TOKEN_KEY, file=TOKEN_FILE_NAME)
         token = out[CONFIG_TOKEN_KEY][0]
@@ -217,10 +218,26 @@ def index():
     if request.args.get('reset'):
         globs.init_globals()
         globs.summary_links = SUMMARY_NAVBAR_PLACEHOLDER
-        globs.reset_btn = None
         app.logger.info("Project data removed")
-    projects, globs.command = parse_config_file()
-    return render_template('index.html', projects=projects, reset_btn=globs.reset_btn, command_btn=globs.command,
+    globs.cc = globs.cc or parse_config_file()
+    read_preferences()
+    missing_projs = globs.cc.list_missing_projects() or None
+    project = request.args.get('path')
+    subproject = request.args.get('sp')
+    app.logger.debug("Selected project:subproject bundle -- {}:{} of types {}:{}"
+                     .format(project, subproject, project.__class__.__name__, subproject.__class__.__name__))
+    with globs.cc as x:
+        if request.args.get('remove'):
+            x.remove_project(path=project, sp=subproject)
+        if request.args.get('populate'):
+            x.populate_project_metadata(paths=project, sp=subproject)
+        if request.args.get('clear'):
+            globs.purge_project_data()
+            x.populate_project_metadata(clear=True)
+    if missing_projs:
+        app.logger.warning("{} projects configs not found: {}".format(len(missing_projs), ", ".join(missing_projs)))
+    app.logger.debug(globs.cc)
+    return render_template('index.html', missing_projects=missing_projs, cc=globs.cc.filter_missing(),
                            selected=globs.selected_project, selected_id=globs.selected_project_id)
 
 
@@ -240,27 +257,31 @@ def set_comp_env():
     if globs.compute_config is None:
         globs.compute_config = divvy.ComputingConfiguration()
     selected_package = request.args.get('compute', type=str)
-    selected_interval = request.args.get('interval', type=int) or globs.poll_interval
-    globs.poll_interval = int(selected_interval)
-    if globs.currently_selected_package is None:
-        globs.currently_selected_package = "default"
+    globs.status_check_interval = int(request.args.get('interval', type=int)
+                                      or globs.status_check_interval
+                                      or POLL_INTERVAL)
+    if globs.compute_package is None:
+        globs.compute_package = "default"
     if selected_package is not None:
         success = globs.compute_config.clean_start(selected_package)
         if not success:
             msg = "Compute package '{}' cannot be activated".format(selected_package)
             app.logger.warning(msg)
             return jsonify(active_settings=render_template('compute_info.html', active_settings=None, msg=msg))
-        globs.currently_selected_package = selected_package
+        globs.compute_package = selected_package
         active_settings = globs.compute_config.get_active_package()
+        write_preferences({"status_check_interval": globs.status_check_interval,
+                       "compute_package": globs.compute_package})
         return jsonify(active_settings=render_template('compute_info.html', active_settings=active_settings))
     active_settings = globs.compute_config.get_active_package()
     notify_not_set = COMPUTE_SETTINGS_VARNAME[0] if \
         globs.compute_config.default_config_file == globs.compute_config.config_file else None
-
+    write_preferences({"status_check_interval": globs.status_check_interval,
+                       "compute_package": globs.compute_package})
     return render_template('preferences.html', env_conf_file=globs.compute_config.config_file,
                            compute_packages=globs.compute_config.list_compute_packages(), active_settings=active_settings,
-                           currently_selected_package=globs.currently_selected_package, notify_not_set=notify_not_set,
-                           default_interval=globs.poll_interval)
+                           compute_package=globs.compute_package, notify_not_set=notify_not_set,
+                           default_interval=globs.status_check_interval)
 
 
 @app.route("/process", methods=['GET', 'POST'])
@@ -268,31 +289,31 @@ def set_comp_env():
 def process():
     from looper import build_parser as blp
     actions = get_positional_args(blp(), sort=True)
-    # this try-except block is used to determine whether the user should be redirected to the index page
-    # to select the project when they land on the process subpage from the set_comp_env subpage
-    if globs.selected_project is None and request.form.get('select_project') is None:
-        app.logger.info("The project is not selected, redirecting to the index page.")
+    try:
+        globs.selected_project, globs.selected_project_id, globs.current_subproj = \
+            select_project(request.form.get('select_project'))
+    except TypeError:
+        app.logger.info("No project selected, redirecting to the index page.")
         flash("No project was selected, choose one from the list below.")
         return redirect(url_for('index'))
-    else:
-        new_selected_project = request.form.get('select_project')
-        if new_selected_project is not None and globs.selected_project != new_selected_project:
-            globs.selected_project = parse_selected_project(new_selected_project)[0]
-            globs.selected_project_id = parse_selected_project(new_selected_project)[1]
-            app.logger.debug("Selected project path: " + globs.selected_project)
-            app.logger.debug("Selected project id: " + globs.selected_project_id)
     config_file = str(os.path.expandvars(os.path.expanduser(globs.selected_project)))
     if globs.p is None:
-        globs.p = Project(config_file)
+        globs.p = Project(config_file, subproject=globs.current_subproj)
     try:
+        # subproject related logic can be removed with the introduction of direct subproject selection in the index
         subprojects = globs.p.subprojects.keys()
     except AttributeError:
         subprojects = None
+    # populating project/subproject metadata and date are treated individually since when the project is activated
+    # we want its subprojects data to be populated but not the dates
+    with globs.cc as x:
+        x.populate_project_metadata(paths=globs.selected_project,
+                                           sp=globs.current_subproj if globs.current_subproj else None)
+        x.project_date(globs.selected_project, globs.current_subproj)
     get_navbar_summary_links()
-    globs.reset_btn = True
     return render_template('process.html', p_info=project_info_dict(globs.p), change=None,
                            selected_subproject=globs.p.subproject, actions=actions, subprojects=subprojects,
-                           interval=globs.poll_interval)
+                           interval=globs.status_check_interval)
 
 
 @app.route('/_background_subproject')
@@ -304,6 +325,8 @@ def background_subproject():
     else:
         globs.p.activate_subproject(sp)
         globs.run = False
+        globs.current_subproj = sp
+        globs.cc.project_date(globs.selected_project, globs.current_subproj)
     globs.summary_requested = None
     get_navbar_summary_links()
     return jsonify(subproj_txt=output, p_info=project_info_dict(globs.p), navbar_links=globs.summary_links)
@@ -370,7 +393,7 @@ def action():
     # establish the looper log path
     # set the selected computing environment in the Project object
     try:
-        globs.p.dcc.activate_package(globs.currently_selected_package)
+        globs.p.dcc.activate_package(globs.compute_package)
     except NameError:
         app.logger.info("The compute package was not selected, using 'default'.")
         globs.p.dcc.activate_package("default")
@@ -387,12 +410,12 @@ def background_check_status():
     if all(not value for value in flags.values()) and not globs.run:
         return jsonify(status_table="No samples were processed yet. " \
                                     "Use <code>looper run</code> and then check the status",
-                       interval=globs.poll_interval)
+                       interval=globs.status_check_interval)
     elif any(value for value in flags.values()):
         return jsonify(status_table=create_status_table(globs.p, final=False) + sample_info_hint(globs.p),
-                       interval=globs.poll_interval)
+                       interval=globs.status_check_interval)
     else:
-        return jsonify(status_table=MISSING_SAMPLE_DATA_TXT, interval=globs.poll_interval)
+        return jsonify(status_table=MISSING_SAMPLE_DATA_TXT, interval=globs.status_check_interval)
 
 
 @app.route('/_background_result')
@@ -423,6 +446,7 @@ def main():
         globs.logging_lvl = logging.DEBUG
     else:
         generate_token(token=parse_token_file())
+    app.logger.setLevel(globs.logging_lvl or logging.INFO)
     app.logger.info("Using python {}".format(python_version()))
     app.run(port=args.port, host='0.0.0.0')