+++ caravel/const.py
@@ -1,9 +1,11 @@
 """ Package constants """
 import os
 from divvy.const import COMPUTE_SETTINGS_VARNAME
-from looper import __version__ as LOOPER_VERSION
+from looper import __version__ as LOOPER_VERSION, Sample as LooperSample
 from peppy import __version__ as PEPPY_VERSION
 from ._version import __version__ as CARAVEL_VERSION
+from ubiquerg import filesize_to_str
+from warnings import warn
 
 
 def get_req_version(module=None):
@@ -34,6 +36,27 @@ def get_req_version(module=None):
         return None
 
 
+def input_sizes(p):
+    """
+    Safely determine all the input file sizes for a project. If any exception raises, return "NA".
+
+    :param looper.Project p: project to determine input file sizes for
+    :return str: sum input file sizes
+    """
+    cumulative_file_size = 0
+    try:
+        for s in p.samples:
+            s = LooperSample(s)
+            for proto in [s.protocol] if isinstance(s.protocol, str) else s.protocol:
+                for piface in p.get_interfaces(proto):
+                    s.set_pipeline_attributes(piface, piface.fetch_pipelines(proto))
+                    cumulative_file_size += s.input_file_size * 1024 ** 3
+        return filesize_to_str(cumulative_file_size)
+    except Exception as e:
+        warn("Could not determine file size for project: '{}'. Got: '{}'".format(p.name, e.__class__.__name__))
+        return "NA"
+
+
 DEFAULT_PORT = 5000
 DEFAULT_LOGGING_LVL = 20
 CONFIG_ENV_VAR = "CARAVEL"
@@ -42,9 +65,8 @@ CONFIG_TOKEN_KEY = "token"
 TOKEN_FILE_NAME = ".caravel_token"
 EXAMPLE_FILENAME = "caravel_demo.yaml"
 TOKEN_LEN = 15
-SET_ELSEWHERE = [["--force-yes"], ["--sp"], ["--compute"], ["--env"], ["--help"], ["--version"],
-                 ["--selector-attribute"], ["--selector-exclude"], ["--selector-include"], ['--resources'],
-                 ['--compute-package'], ['']]
+SET_ELSEWHERE = [["--force-yes"], ["--sp"], ["--env"], ["--help"], ["--version"], ["--selector-attribute"],
+                 ["--selector-exclude"], ["--selector-include"], ['--resources'], ['--compute-package'], ['']]
 LOG_FILENAME = "caravel.log"
 REQUIRED_LOOPER_VERSION = get_req_version("loopercli")["loopercli"]
 REQUIRED_PEPPY_VERSION = get_req_version("peppy")["peppy"]
@@ -61,4 +83,43 @@ MISSING_SAMPLE_DATA_TXT = "<code>looper run</code> was called, but not all the s
                             "</br>Possible reasons: <ul style='padding-left: 30px;'>"\
                             "<li>all jobs are still in a queue</li>" \
                             "<li>submission was not successful</li></ul>"
+REQ_CFG_VERSION = 0.2
+# this preferences/types can be set in the config file under "preferences" key
+PREFERENCES_NAMES_TYPES = {"status_check_interval": int,
+                           "compute_package": str}
+# mapping of looper.Project metadata of interest and lambda expressions extracting them
+PROJECT_MDATA_FUN = {"name": lambda p: p.name,
+                     "names_sp": lambda p: ", ".join(p.subprojects.keys()),
+                     "num_sp": lambda p: len(p.subprojects.keys()),
+                     "num_samples": lambda p: p.num_samples,
+                     "protocols": lambda p: ", ".join(sorted(p.protocols)),
+                     "inputs_size": input_sizes}
 
+
+"""
+Config file structure determination 
+"""
+# config file structure related consts
+
+CFG_NAME = "caravel configuration"
+CFG_ENV_VARS = ["CARAVEL"]
+
+CFG_VERSION_KEY = "config_version"
+CFG_PROJECTS_KEY = "projects"
+CFG_SUBPROJECTS_KEY = "subprojects"
+CFG_PREFERENCES_KEY = "preferences"
+CFG_PROJECT_NAME_KEY = "name"
+CFG_PROJECT_DESC_KEY = "project_description"
+
+CFG_EXAMPLE = """
+# example {cfg_name} structure
+{version}: 0.2
+{preferences}:
+  status_check_interval: 2
+
+{projects}:
+  /home/johndoe/projects/config_rnaseq.yaml:
+    {name}: rnaseq_february
+    {desc}: Processes a set of RNA-seq samples 
+""".format(cfg_name=CFG_NAME, version=CFG_VERSION_KEY, projects=CFG_PROJECTS_KEY, preferences=CFG_PREFERENCES_KEY,
+           name=CFG_PROJECT_NAME_KEY, desc=CFG_PROJECT_DESC_KEY)