+++ caravel/globs.py
@@ -1,6 +1,3 @@
-from .const import *
-
-
 def init_globals():
     """
     This function initializes global variables, which then can be used in the whole app
@@ -17,13 +14,13 @@ def init_globals():
     global logging_lvl
     global summary_links
     global dests
-    global reset_btn
     global command
-    global currently_selected_package
+    global compute_package
     global current_subproj
     global summary_requested
     global run
-    global poll_interval
+    global status_check_interval
+    global cc
 
     summarizer = None
     p = None
@@ -32,15 +29,47 @@ def init_globals():
     log_path = None
     act = None
     compute_config = None
-    logging_lvl = DEFAULT_LOGGING_LVL
+    logging_lvl = None
     summary_links = None
     dests = None
-    reset_btn = None
     command = None
-    currently_selected_package = None
+    compute_package = None
     current_subproj = None
     summary_requested = None
     run = None
-    poll_interval = POLL_INTERVAL
+    status_check_interval = None
+    cc = None
+
+
+def purge_project_data():
+    """
+    Removes project related data while preserving settings/information unassociated with the currently selected project.
+    Can be used before new project/subproject initialization.
+    """
+    global summarizer
+    global p
+    global selected_project
+    global selected_project_id
+    global log_path
+    global act
+    global summary_links
+    global dests
+    global command
+    global current_subproj
+    global summary_requested
+    global run
+
+    summarizer = None
+    p = None
+    selected_project = None
+    selected_project_id = None
+    log_path = None
+    act = None
+    summary_links = None
+    dests = None
+    command = None
+    current_subproj = None
+    summary_requested = None
+    run = None